floattt

+++ headers/event.hpp
@@ -8,17 +8,17 @@ class Event
 protected:
     pair<int, int> pos;
     string path;
-    int scale;
+    float scale;
 
     virtual void actiune() = 0;
-    Event(const pair<int, int> &, const string &, const int);
+    Event(const pair<int, int> &, const string &, const float);
 
 public:
     virtual ~Event() = default;
     Event() = default;
     pair<int, int> get_pos() const;
     string get_path() const;
-    int get_scale() const;
+    float get_scale() const;
 };
 
 class Fruit : public Event
@@ -29,7 +29,7 @@ class Fruit : public Event
 public:
     virtual ~Fruit() = default;
     Fruit() = default;
-    Fruit(const pair<int, int> &, const string &, const int, const int);
+    Fruit(const pair<int, int> &, const string &, const float, const int);
 };
 
 class Flower : public Event
@@ -42,7 +42,7 @@ class Flower : public Event
 public:
     virtual ~Flower() = default;
     Flower() = default;
-    Flower(const pair<int, int> &, const string &, const int, const int);
+    Flower(const pair<int, int> &, const string &, const float, const int);
 };
 
 class Key : public Event
@@ -52,7 +52,7 @@ public:
     virtual ~Key() = default;
     void actiune() override{};
     Key() = default;
-    Key(const pair<int, int> &, const string &, const int);
+    Key(const pair<int, int> &, const string &, const float);
 };
 
 #endif
